Fix broken texture name
As far as I can tell, until about 15 minutes ago, this would have caused
HUDReplacer to crash if it ever actually tried to replace a texture
called `axis_preset_square`. Luckily the KAL textures are loaded
on-demand so it likely hasn't been a big issue.

This commit fixes the texture name so that it is no longer an issue.

With the new version this texture will just not be loaded and a error
message will be put in the logs.