fix: Move VAB-specific override into a sceneSpecific folder
There are two ambiguous texture replacements in ZTheme:
- ZTheme/PluginData/VAB/panel4_5c626e_d_r124#24x24.png
- ZTheme/PluginData/UI/panel4_5c626e_d_r124.png

Before v1.3.1 HUDReplacer would pick whichever of these happened to come
last in the dictionary it used to store the replacement paths. This
usually happened to result in the UI variant being chosen.

In v1.3.2 and later (v1.3.1 was bugged here) HUDReplacer now correctly
picks the VAB version as it has a higher priority. This means that the
sidebar in R&D ends up being darker because it is using the VAB texture.

This commit is my best attempt at fixing what seems to be intended here.
If you just want to keep the existing behaviour then you should just
delete the VAB texture here instead.